+++ app/auth/oauth2.py
@@ -6,8 +6,7 @@ from fastapi.security import HTTPBearer
 
 from app.auth.jwt_token import decode_token
 from app.core.exceptions import ForbiddenException, Unauthorized
-from app.schemas.auth import User
-from app.schemas.workplace import Role, UserAssignedWorkplace
+from app.schemas.documents import Role, User, UserAssignedWorkplace
 
 oauth2_scheme = HTTPBearer()
