+++ app/routers/auth.py
@@ -8,11 +8,12 @@ from app.auth.hash import get_password_hash, verify_password
 from app.auth.jwt_token import create_access_token, create_refresh_token
 from app.auth.oauth2 import get_current_user
 from app.config import client_api_settings
-from app.core import UserRegister
 from app.core.email import Email
 from app.core.exceptions import EmailVerificationException, UserFoundException
 from app.core.redis_session import Redis
-from app.schemas.auth import SuccessfulResponse, Token, TokenData, User
+from app.schemas.documents import User
+from app.schemas.models.auth import Token, TokenData, UserRegister
+from app.schemas.models.models import SuccessfulResponse
 
 router = APIRouter(tags=["Auth"])
