+++ app/schemas/__init__.py
@@ -1,7 +1,4 @@
-from .auth import User
-from .issue import Issue
-from .sprint import Sprint
-from .workplace import UserAssignedWorkplace, Workplace
+from .documents import Issue, Sprint, User, UserAssignedWorkplace, Workplace
 
 __beanie_models__ = [User, Workplace, Sprint, UserAssignedWorkplace, Issue]
